reorganise repository
